scripts/requirements.txt
requests~=2.31.0
openai~=1.14.0
python-dotenv~=1.0.1
playwright~=1.42.0
lark-oapi~=1.2.9
fastapi~=0.109.0
uvicorn~=0.27.0
loguru~=0.7.2
curl_cffi~=0.6.2
